+++ CHANGELOG.md
@@ -33,7 +33,7 @@ All notable changes to this project will be documented in this file.
 - `SelectAlbum` queue keybind
 - More felixbility to `Browser` panes. They can now be configured to any arbitrary levels of nesting
   instead of hardcoded `["root_tag", "album"]`
-- Added `RecentlyPlayed` browser pane showing songs ordered by the `lastPlayed` sticker. The sticker
+- Added `Sticker` browser pane showing songs that have the selected sticker. The sticker
   needs to be tracked externally by for example [rmpcd plugin](https://github.com/rmpc-org/rmpcd-lastplayed).
 - Added `zt`, `zz` and `zb` Vim-style navigation keybinds to scroll the focused entry to the top,
   middle or bottom of the viewport